Get the Proof (Merkle Tree) for the hash calculated from an R object or file.

Wrapper around https://api.originstamp.com/swagger/swagger-ui.html#/proof/getProof. The function downloads the merkle tree as proof.

get_proof(x, file, error_on_fail = TRUE, url = api_url(), key = api_key())

Arguments

x
an R object of which a hash will be calculated using the function hash(x). The resulting hash will be submitted to OriginStamp.

file
if provided, file name to store the merkle tree as xml file to.

error_on_fail
if TRUE, raise error when api call fails, otherwise return the failed response.

url
the url of the api. The default is to use the url as returned by the function api_url()

key
the api key. The default is to use the key as returned by the function api_key()

Value

object of class OriginStampResponse, with an additional element, return$proof, which contains the merkle tree as an xml_document.

Details

The behavior depends on the class of the argument x:

an object of class hash as returned by the package openssl: the hash is submitted to OriginStamp

character vector of length 1 containing the name of an existing file: the hash of the file is calculated and submitted to OriginStamp

any other R object: the hash is calculated using the function hash() and submitted to OriginStamp

Examples

if (FALSE) {
# Retrieve complete merkle tree proof
x <- "2c5d36be542f8f0e7345d77753a5d7ea61a443ba6a9a86bb060332ad56dba38e"
class(x) <- "hash"
get_proof(
  x = x
)
}