Calculate sha256 hash

A generic function function to calculate the sha256 hash from an R object or from a file.

hash(x)

Arguments

x
object

Value

sha256 hash object

Details

The behavior depends on the class of the argument x:

an object of class sha256 as returned by the package openssl: the hash is returned as is

character vector of length 1 containing the name of an existing file: the sha256 hash of the file is calculated

any other R object: is serialized using `serialize(x, connection = NULL, ascii = FALSE, xdr = TRUE, version = 2, refhook = NULL)`` and the sha256 hash is created from that serialized object.

The methods doing the work are:

hash.hash(x): returns the object x

hash.file(): assumes x is an existing file and tries to calculate the hash from the file

hash.default(): calculates the hash from the serialized R object x NB: These functions do assume the object x is what they expect it to be. It is recommended to use the generic function hash() as this functiuon auti=omatically selects the most appropriate method. Only use these if you need to force a certain method (e.g. calculate the hash of a file name instead of the existing file).

Remark: if a hash is in haracter form, it can be converted to a hash object, by assigning the class hash (see example).

Examples

## hash of a file:

hash(system.file("DESCRIPTION", package = "ROriginStamp"))

#> 
#> Create sha356 hash from R file x [/private/var/folders/50/wcr5bjwn75q595n6x82gxj280000gq/T/RtmpKinAfk/temp_libpatha740440e8263/ROriginStamp/DESCRIPTION]

#> sha256 2b:73:f5:01:16:eb:d2:9b:c9:c9:d5:20:7b:72:29:c3:b5:d2:9a:09:10:e6:c1:3f:b1:90:5b:9b:ac:e6:77:7a 

## convert a character vector to a hash:

x <- "2c5d36be542f8f0e7345d77753a5d7ea61a443ba6a9a86bb060332ad56dba38e"
class(x) <- "hash"
hash(x)

#> Error in UseMethod("hash", x): no applicable method for 'hash' applied to an object of class "hash"